Revert tmp rename
